Reorganise based on review feedback
