+++ src/containers/Elements/index.js
@@ -1,2 +1,3 @@
-export { default as NodeSuggester } from './NodeSuggester';
+export { default as NodeProvider } from './NodeProvider';
+export { default as NodeProviderPanels } from './NodeProviderPanels';
 export { default as PromptSwiper } from './PromptSwiper';
